Graph Editor: Preserve active curve when using AKEY to toggle selection status of keyframe verts
Previously, every time you toggled the selection of all keyframes (using AKEY),
the active curve would get deselected and deactivated. However, this was a pain
when trying to tweak the shape of a particular curve, as doing this would cause
that curve to either fade into the background or into the jumble of other
curves.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@59756 954f8c5b-7b00-dc11-b283-0030488c597c