Film response curves implemented as a looks
This commit implement's OCIO's Looks idea which
is about applying some color correction on the
buffer before it get's affected by a display
transform.

This is mainly used to modify images in an
artistics way.

Currently we've got looks generated from film
response curves for all sorts of cameras.

Patch by both of me and Brecht.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@59945 954f8c5b-7b00-dc11-b283-0030488c597c