Sculpting:
Modify calculation of sculpt plane to only take into account forward
facing vertices. This will solve cases where sculpting on a volume with
the two sides of the mesh inside the brush radius could move the sculpt
plane inside the mesh volume.

To completely fix the issue where the mesh would "stick" the two sides
of the mesh together on the sculpt plane (for instance for clay strip
brushes), user should enable "front face only". Perhaps some brushes,
like clay strips should enforce this and not present the option in the
first place.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@60121 954f8c5b-7b00-dc11-b283-0030488c597c