Derivative map baker
Added support for derivative map baking, which
is accessable as a dedicated baker type. Works
pretty much the same as displacement map baker,
but gives you derivative map.

In fact, inernally this baker is just a filter
which applies on the result of displacement map.

Both regular and multires baking are supported.

Patch by Morten Mikkelsen and self.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@60631 954f8c5b-7b00-dc11-b283-0030488c597c