Fix #37090: Blender crashes on "Make single user" option
issue was caused by id->mewid pointing to an invalid memory
after file load.

The rule here: all the tools need to reset it to NULL (or
other value they need) in the beginning.

Currently some tools are doing newid clear in the beginning
and some does it in the end. We need to clean it up so clear
only happens in the beginning.

But ideal we need some kind NewIDContext to make duplication
safe for threading.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@60784 954f8c5b-7b00-dc11-b283-0030488c597c